api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic>=2.13.2",
    "pyjwt>=2.12.1",
    "redis>=7.4.0",
    "uvicorn[standard]>=0.44.0",
]
